two passing tests and a helper spec
