+++ spec/tddetective-spec.js
@@ -1,6 +1,7 @@
 'use babel';
 
 import Tddetective from '../lib/tddetective';
+import HelperModule from './helper-spec';
 
 // Use the command `window:run-package-specs` (cmd-alt-ctrl-p) to run specs.
 //
@@ -13,11 +14,14 @@ describe('Tddetective', () => {
   beforeEach(() => {
     workspaceElement = atom.views.getView(atom.workspace);
     activationPromise = atom.packages.activatePackage('tddetective');
+
+    spyOn(Tddetective, "listenToChanges").andCallFake(function(){
+
+    })
   });
 
-  describe("listening function", () => {
-    it("the listening function is called in activate function", () => {
-      spyOn(Tddetective, 'listenToChanges');
+  describe("listening functionality", () => {
+    it("listenToChanges is called by toggle command", () => {
       atom.commands.dispatch(workspaceElement, 'tddetective:toggle');
 
       waitsForPromise(() => {
@@ -29,8 +33,8 @@ describe('Tddetective', () => {
       })
     })
 
-    it("listens on changes and calls findClassNamesAndCallToggle", () => {
-      spyOn(Tddetective, 'findClassNamesAndCallToggle').andCallThrough();
+    it("listens on changes and calls pickUpChanges", () => {
+      spyOn(Tddetective, '_pickUpChanges').andCallThrough();
       atom.commands.dispatch(workspaceElement, 'tddetective:toggle');
 
       waitsForPromise(() => {
@@ -38,7 +42,7 @@ describe('Tddetective', () => {
       });
 
       runs(() => {
-        expect(Tddetective.findClassNamesAndCallToggle).toHaveBeenCalled();
+        expect(Tddetective._pickUpChanges).toHaveBeenCalled();
       })
     })
 
@@ -63,23 +67,13 @@ describe('Tddetective', () => {
     })
 
     xit("find the class names in the active text editor", () => {
-      var testString = "class Bike"
 
-      var dir = atom.config.configDirPath + "/packages/TDDetective/spec/tddetective-spec.rb"
-      var editor = atom.workspace.buildTextEditor();
-      editor.buffer.cachedDiskContents = testString
-      editor.buffer.cachedText = testString
-      console.log(dir)
-      editor.buffer.file = new File([], false)
-      editor.buffer.file.cachedContents = "class Bike"
-      console.log(editor)
-
-      var buffer = editor.getBuffer()
-      var classNames = []
-
-
-      console.log(buffer)
-      console.log(classNames)
+      helperModule.createMockEditor()
+      helperModule.addDataToMockEditor()
+        //
+        // var buffer = editor.getBuffer()
+        // var classNames = []
+        //
       var selection = Tddetective.findClassNames(editor);
       expect(selection.includes("Bike")).toEqual(true)
 
@@ -88,46 +82,9 @@ describe('Tddetective', () => {
   })
 
 
-  describe('when the tddetective:toggle event is triggered', () => {
-    xit('hides and shows the modal panel', () => {
-      // Before the activation event the view is not on the DOM, and no panel
-      // has been created
-      expect(workspaceElement.querySelector('.tddetective')).not.toExist();
-
-      // This is an activation event, triggering it will cause the package to be
-      // activated.
-      atom.commands.dispatch(workspaceElement, 'tddetective:toggle');
-
-      waitsForPromise(() => {
-        return activationPromise;
-      });
-
-      runs(() => {
-        expect(workspaceElement.querySelector('.tddetective')).toExist();
-
-        let tddetectiveElement = workspaceElement.querySelector('.tddetective');
-        expect(tddetectiveElement).toExist();
-
-        let tddetectivePanel = atom.workspace.panelForItem(tddetectiveElement);
-        expect(tddetectivePanel.isVisible()).toBe(true);
-        atom.commands.dispatch(workspaceElement, 'tddetective:toggle');
-        expect(tddetectivePanel.isVisible()).toBe(false);
-      });
-    });
-
-    it('checks whether aClassName is a filename in spec dir', function(){
-
-      let tddetectiveElement = workspaceElement.querySelector('.tddetective');
-    })
+  describe('CLASS NAME FIND', () => {
 
     it('checks whether aClassName is a filename in spec dir', () => {
-      // This test shows you an integration test testing at the view level.
-
-      // Attaching the workspaceElement to the DOM is required to allow the
-      // `toBeVisible()` matchers to work. Anything testing visibility or focus
-      // requires that the workspaceElement is on the DOM. Tests that attach the
-      // workspaceElement to the DOM are generally slower than those off DOM.
-
       // This is an activation event, triggering it causes the package to be
       // activated.
       atom.commands.dispatch(workspaceElement, 'tddetective:toggle');
@@ -139,9 +96,10 @@ describe('Tddetective', () => {
       runs(() => {
         // let tddetectiveElement = workspaceElement.querySelector('.tddetective');
         var dir = atom.config.configDirPath + "/packages/TDDetective/spec" ;
-        spyOn(Tddetective, 'getSpecPath').andReturn(dir);
+        spyOn(Tddetective, '_getSpecPath').andReturn(dir);
+        console.log(dir)
 
-        expect(Tddetective.hasSpecFileName("tddetective")).toEqual(true)
+        expect(Tddetective._hasSpecFileName("tddetective")).toEqual(true)
 
       });
     });
